+++ src/test/java/ru/netology/stats/StatServiceTest.java
@@ -26,7 +26,7 @@ public class StatServiceTest {
     @Test
     void statMaxMonth() {
         StatService service = new StatService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 8;
         long actual = service.maxSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -35,7 +35,7 @@ public class StatServiceTest {
     @Test
     void statMinMonth() {
         StatService service = new StatService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 9;
         long actual = service.minSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -46,7 +46,7 @@ public class StatServiceTest {
         StatService service = new StatService();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
-        long actual = service.salesBelowAverage(sales);
+        int actual = service.salesBelowAverage(sales);
         Assertions.assertEquals(expected, actual);
     }
